requirements.txt
python-ceilometerclient
python-keystoneclient
iso8601
oslo.config>=1.2.0
pbr>=0.6,<1.0
